First python script to create a gender based model
The script creates a simple gender based model in python 3 (NB!.. not
2.7).

Note that tinus_titanic1 is my own code. The other python files were
are from the tutorial. They only run in python 2.7!

train.csv is the training data set for the machine learning problem. It
includes a list of all the python

test.csv is the testing data set for the problem. You need to submit a file
with only 2 columns, passenger id and survived.